HV:treewide:Cleanup the type for parameters of bitmap  operations
For reducing sign conversion in hypervisor:
Update parameters of bitmap operations as unsigned type;
Update the input of related caller as unsigned type when the
caller's input parameter is const variable or the variable is
only used by bitmap operations.

V1-->V2:
        (1) Explicit casting for the first parameter
            of all bitmap operations;
        (2) Remove mask operation for explicit casting
            of all bitmap operations, since masking is
            useless. Otherwise, this trucation is dangerous.
V2-->V3:
        (1) Explicit casting for all bitmap operations parameter;
        (2) Masking bit offset with 6-bit;
        (3) Add few comments about bit offset.
V3-->V4:
        add '\' for some statement of bitmap macro

Acked-by: Eddie Dong <eddie.dong@intel.com>

+++ hypervisor/arch/x86/guest/vcpu.c
@@ -387,7 +387,7 @@ int prepare_vcpu(struct vm *vm, uint16_t pcpu_id)
 	return ret;
 }
 
-void request_vcpu_pre_work(struct vcpu *vcpu, int pre_work_id)
+void request_vcpu_pre_work(struct vcpu *vcpu, uint16_t pre_work_id)
 {
 	bitmap_set(pre_work_id, &vcpu->pending_pre_work);
 }
